training_prediction/requirements.txt
pandas
scikit-learn
joblib
sqlalchemy
psycopg2-binary
